# Conflicts:
#   src/Entity/Annee.php

+++ src/Controller/UserController.php
@@ -2,20 +2,13 @@
 
 namespace App\Controller;
 
-use ApiPlatform\Core\Validator\ValidatorInterface;
 use App\Entity\User;
 use App\Form\UserType;
 use App\Repository\UserRepository;
-use Doctrine\ORM\EntityManager;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
-use Symfony\Component\Serializer\Exception\NotEncodableValueException;
-use Symfony\Component\Serializer\Normalizer\NormalizableInterface;
-use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
-use Symfony\Component\Serializer\SerializerInterface;
 
 /**
  * @Route("/user")
@@ -25,49 +18,20 @@ class UserController extends AbstractController
     /**
      * @Route("/", name="user_index", methods={"GET"})
      */
-    public function index(UserRepository $userRepository)
+    public function index(UserRepository $userRepository): Response
     {
-        $users = $userRepository->findAll();
-        //$userNormalises = $normalizer->serialize($users,"json", ['groups'=>'user:read']);
-        //$json = json_encode($userNormalises);
-        //$json = $normalizer->serialize($users,"json", ['groups'=>'user:read']);
-
-
-        /*$response = new Response($json, 200, [
-            "Content-Type" => "application/json"
-        ] );*/
-        //$response= new JsonResponse($json,200,[],true);
-        $response=$this->json($users, 200,[], ["groups"=>"user:read"]);
-
-        return $response;
+        return $this->render('user/index.html.twig', [
+            'users' => $userRepository->findAll(),
+        ]);
     }
 
     /**
-     * @Route("/new", name="user_new", methods={"POST"})
+     * @Route("/new", name="user_new", methods={"GET","POST"})
      */
-    public function new(Request $request, SerializerInterface $serializer, EntityManager $manager,
-    ValidatorInterface $validator): Response
+    public function new(Request $request): Response
     {
         $user = new User();
-        $jsonRecu = $request->getContent();
-        try {
-            $users = $serializer ->deserialize($jsonRecu, User::class, 'json');
-            $errors = $validator->validate($users);
-
-            if (count($errors) > 0){
-                return $this->json($errors, 400);
-            }
-            $manager->persist($users);
-            $manager->flush();
-            return $this->json($users, 201, [], ["groups"=>"user:read"]);
-        }catch (NotEncodableValueException $e){
-            return $this->json([
-                "status"=>400,
-                "message"=>$e->getMessage()
-            ],400);
-        }
-
-        /*$form = $this->createForm(UserType::class, $user);
+        $form = $this->createForm(UserType::class, $user);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
@@ -81,7 +45,7 @@ class UserController extends AbstractController
         return $this->render('user/new.html.twig', [
             'user' => $user,
             'form' => $form->createView(),
-        ]);*/
+        ]);
     }
 
     /**
@@ -89,31 +53,17 @@ class UserController extends AbstractController
      */
     public function show(User $user): Response
     {
-        $response=$this->json($user, 200,[], ["groups"=>"user:read"]);
-
-        return $response;
-        /*return $this->render('user/show.html.twig', [
+        return $this->render('user/show.html.twig', [
             'user' => $user,
-        ]);*/
+        ]);
     }
 
     /**
-     * @Route("/{id}/edit", name="user_edit", methods={"PUT"})
+     * @Route("/{id}/edit", name="user_edit", methods={"GET","POST"})
      */
-    public function edit(Request $request, User $user,SerializerInterface $serializer, EntityManager $manager, ValidatorInterface $validator): Response
+    public function edit(Request $request, User $user): Response
     {
-        $jsonRecu = $request->getContent();
-        $users = $serializer ->deserialize($jsonRecu, User::class, 'json');
-        $errors = $validator->validate($users);
-
-        if (count($errors) > 0){
-            return $this->json($errors, 400);
-        }
-
-        $manager->flush();
-        return $this->json($users, 202, [], ["groups"=>"user:read"]);
-
-        /*$form = $this->createForm(UserType::class, $user);
+        $form = $this->createForm(UserType::class, $user);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
@@ -125,7 +75,7 @@ class UserController extends AbstractController
         return $this->render('user/edit.html.twig', [
             'user' => $user,
             'form' => $form->createView(),
-        ]);*/
+        ]);
     }
 
     /**
@@ -133,19 +83,12 @@ class UserController extends AbstractController
      */
     public function delete(Request $request, User $user): Response
     {
-
-        $response=$this->json($user, 200,[], ["groups"=>"user:read"]);
-        $entityManager = $this->getDoctrine()->getManager();
-        $entityManager->remove($user);
-        $entityManager->flush();
-        return $response;
-
-        /*if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
+        if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($user);
             $entityManager->flush();
         }
 
-        return $this->redirectToRoute('user_index');*/
+        return $this->redirectToRoute('user_index');
     }
 }
